requirements.txt
asgiref==3.2.7
Django==3.0.7
django-debug-toolbar==2.2
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
django_heroku